# Conflicts:
#	kjpt-platform/src/main/webapp/layuiadmin/js/common.js

+++ kjpt-platform/src/main/webapp/kjpt/achieve/achieve_listJS.js
@@ -36,7 +36,6 @@ layui.use(['table', 'form','laydate'], function() {
     //开始日期
     var insStart = laydate.render({
         elem: '#inputStart'
-        ,min: 0
         ,done: function(value, date){
             //更新结束日期的最小日期
             insEnd.config.min = lay.extend({}, date, {
@@ -65,26 +64,40 @@ layui.use(['table', 'form','laydate'], function() {
     return false;
   });
   $('[lay-filter="formSubmit"]').click();
+    $("#reset").click(function () {
 
+        queryTable('')
+    })
   // 删除申请
   $('#delItem').on('click', function() {
     // 获取被选中的行数据
     var activeData = table.checkStatus('tableDemo').data;
     if (activeData.length) {
-        httpModule({
-            url: "/achieve-api/delete/"+activeData[0].id,
-            type: 'DELETE',
-            success: function(relData) {
-                if(relData.code==0){
-                    layer.msg('删除成功!', {icon: 1});
-                    top.layer.closeAll(); // 关闭弹窗
-                    queryTable('')
+        if(activeData[0].auditStatus==0||activeData[0].auditStatus==3){
+            httpModule({
+                url: "/achieve-api/delete/"+activeData[0].id,
+                type: 'DELETE',
+                success: function(relData) {
+                    if(relData.code==0){
+                        layer.msg('删除成功!', {icon: 1});
+                        top.layer.closeAll(); // 关闭弹窗
+                        queryTable('')
+                    }
                 }
-            }
-        });
+            });
+        }else {
+            top.layer.msg('当前申请状态不能删除！');
+        }
+
     } else {
       top.layer.msg('请选择需要删除的数据！');
     }
+  })
+  
+  //流程
+  $('#flow').on('click', function() {
+	    var activeData = table.checkStatus('tableDemo').data;
+		dealFlow(activeData[0].id);
   })
   
   // 新增、编辑、查看
